move calculateSortOrder tests to controllers.listsController.test.ts
